requirements.txt
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.4.2
Django==2.0.1
djangorestframework==3.7.7
feedparser==5.2.1
gunicorn==19.7.1
idna==2.6
psycopg2==2.7.3.2
python-twitch-client==0.4.0
pytz==2017.3
requests==2.18.4
simplejson==3.13.2
six==1.11.0
ts3==1.0.6
urllib3==1.22
whitenoise==3.3.1
